requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
alembic==1.13.2
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
itsdangerous==2.2.0
APScheduler==3.10.4
python-dotenv==1.0.1
